refactor(query): refactor code struct
