+++ src/query/service/src/pipelines/processors/transforms/new_hash_join/transform_hash_join.rs
@@ -29,12 +29,12 @@ use tokio::sync::Barrier;
 
 use crate::pipelines::processors::transforms::new_hash_join::join::Join;
 use crate::pipelines::processors::transforms::new_hash_join::join::JoinStream;
+use crate::pipelines::processors::transforms::new_hash_join::performance::PerformanceContext;
 use crate::pipelines::processors::transforms::new_hash_join::runtime_filter::PlanRuntimeFilterDesc;
 
 pub struct TransformHashJoin {
     build_port: Arc<InputPort>,
     probe_port: Arc<InputPort>,
-
     joined_port: Arc<OutputPort>,
 
     stage: Stage,
@@ -42,7 +42,6 @@ pub struct TransformHashJoin {
     joined_data: Option<DataBlock>,
     stage_sync_barrier: Arc<Barrier>,
     projection: ColumnSet,
-    initialize: bool,
     rf_desc: Arc<PlanRuntimeFilterDesc>,
 }
 
@@ -61,15 +60,14 @@ impl TransformHashJoin {
             probe_port,
             joined_port,
             join,
-            joined_data: None,
-            stage_sync_barrier,
+            rf_desc,
             projection,
-            initialize: false,
+            stage_sync_barrier,
+            joined_data: None,
             stage: Stage::Build(BuildState {
                 finished: false,
                 build_data: None,
             }),
-            rf_desc,
         }))
     }
 }
@@ -112,21 +110,11 @@ impl Processor for TransformHashJoin {
         }
 
         match &mut self.stage {
-            Stage::Build(state) => match state.event(&self.build_port)? {
-                Event::NeedData if !self.initialize => {
-                    self.initialize = true;
-                    // self.probe_port.set_need_data();
-                    Ok(Event::NeedData)
-                }
-                other => Ok(other),
-            },
+            Stage::Build(state) => state.event(&self.build_port),
             Stage::BuildFinal(state) => state.event(),
             Stage::Probe(state) => state.event(&self.probe_port),
             Stage::ProbeFinal(state) => state.event(&self.joined_port),
-            Stage::Finished => {
-                self.joined_port.finish();
-                Ok(Event::Finished)
-            }
+            Stage::Finished => Ok(Event::Finished),
         }
     }
 
@@ -155,7 +143,7 @@ impl Processor for TransformHashJoin {
             Stage::Probe(state) => {
                 if let Some(probe_data) = state.input_data.take() {
                     let stream = self.join.probe_block(probe_data)?;
-                    state.stream = Some(stream);
+                    state.stream = Some(unsafe { std::mem::transmute(stream) });
                 }
 
                 if let Some(mut stream) = state.stream.take() {
@@ -170,7 +158,8 @@ impl Processor for TransformHashJoin {
             Stage::ProbeFinal(state) => {
                 if !state.initialized {
                     state.initialized = true;
-                    state.stream = Some(self.join.final_probe()?);
+                    let final_stream = self.join.final_probe()?;
+                    state.stream = Some(unsafe { std::mem::transmute(final_stream) });
                 }
 
                 if let Some(mut stream) = state.stream.take() {
